Implement Theia save feature for GLSP client
This change fixes #36 by enhancing the GLSP client to support Theia's
saveable protocol. On the server, I had to add the @Singleton annotation
to enforce the use of the same model factory instance across all action
handlers that require it. This however is rather a workaround for now,
as we'll need a mechanism to support multiple servers in the long run
anyway.