Introduce tool manager to support different editing modes
With the previous architecture, we could only register tools (key and
mouse listeners) in a way that they were always active at the same time.
This made it difficult to implement a palette because a palette
essentially activates and deactivates different editing tools which
implement different editing behavior. So these tools had to be aware of
a widget-global state to check whether they should do something or not,
and this also opened up a source of issues, where these tools may
interfere with each others' behavior unexpectedly.

Therefore, we introduce a widget-global tool manager which manages the
state of tools, being able to enable and disable them. The tool manager
itself is controlled by sending actions to the action dispatcher.

The tool manager supports two types of tools, standard tools and "other"
tools. Standard tools become active in the common editing state until
other tools are explicitly activated. Once a tool is finished, it can
send an action to enable standard tools again. Note that this nicely
enables us to support a "stamp mode" where a tool remains active when
e.g. Ctrl is pressed while using it.

Signed-off-by: Philip Langer <planger@eclipsesource.com>