Prevent node creation on mouse up events after dragging (#167)
Introduces `DragAwareMouseListener` to make it easier for mouse
listeners based tools to react only to mouse up events without prior
dragging or with prior dragging.

The node creation tool as well as the edge creation tool now only react
to non-dragging mouse up events, so that the user can drag in the
diagram without changing the state of the node or edge creation tools.

This fixes #167.