Fix #185: Ensure elements can be resized after using the reconnect tool
The selected reconnect handles mess up the selection tracking of the
change bounds tool, as they never get properly de-selected through mouse
or key events (how the change bounds tool tracks selection). This fix
therefore ensures that reconnect handles are never selected in the first
place.

As a result, the connectable that you click when changing the source or
target of an edge, now shows resize handles (as both are default
tools and always active).