Fix #196 by enforcing semicolons and line at eof
Unfortunately, I didn't get this to work with the current TSLint VSCode
plug-in but only with the one that we've used before. So I changed the
plugin also in the extensions.json.

Sorry this is a very large PR which likely leads to conflicts as it fixes all
the errors and warnings caused by the new enforced tslint rules.