Suggestion for config improvements
This change switches from the depricated eg2.tslint extension
to the now current ms-vscode.vscode-typescript-tslint-plugin.
Moreover, this fixes the configuration that errors and warnings
would show up in the editor and that quick-fix would work
reliably on save.

Moreover, this omits the usage of import-splitnsort and relies
on typescript-hero only. I feel that this is the more stable option
for now, but this is open for discussion.