Improve collecting dependent elements for element deletions
In case of deleting edges, the model index got unnecessarily polluted due to calling the operation SModelIndex.get[Incomging|Outgoing]Edges also for edges when computing dependent elements that should be deleted together with the edges. This commit introduces a condition so that incoming and outgoing edges are only considered as dependent elements of nodes, i.e., the mentioned SModelIndex operations are only for nodes.