Add action for clearing markers
This commit introduces the Clear Markes Action and corresponding command, which are responsible for clearing old markers visualized on the diagram. This can be necessary in case of a re-validation, namely when previously retrieved markers become obsolete. Also, it can be used to allow the user to clear all visualized makers.