Add handlers for directly writing into XMI resources via model access
This change enhances the example server for dealing with EMF XMI models
directly by introducing handlers that read from and directly modify the
EMF resources. After each change, the graph factory is retriggered so
that a new GModel is generated out of the EMF resources to reflect the
applied changes.

With this strategy, we avoid a bidirectional synchronization between EMF
models and GModel but rather create a circle in which we translate the
EMF resources into a GModel and handlers only translate changes applied
in the editor to the underlying EMF models directly, which again can be
re-translated into a GModel with the same factory as above.

The access to the EMF models is provided via a model access that is kept
in a specialized graphical model state to avoid reloading the resources
again and again. Once hooked up with the model server, this model access
can be the sole point of communication with the model server.