Add support for context menu and harmonize with command palette
This change introduces client-side support for context menus including
an implementation for Theia, as well as a server-side provider
mechanism for context menu items.
This mechanism has been harmonized with the provider for command palette
actions so that there is only one "context commands" request/set pair.
The server as well as the client framework separates them, however, to
make it clearer for implementers.

Alongside this new feature, this change also applies a few of cleanups.